+++ crypto/s2n_stream_cipher_rc4.c
@@ -30,10 +30,10 @@ static int s2n_stream_cipher_rc4_encrypt(struct s2n_session_key *key, struct s2n
 {
     gte_check(out->size, in->size);
 
-    int len = out->size;
+    int len = (int)out->size;
     GUARD_OSSL(EVP_EncryptUpdate(key->evp_cipher_ctx, out->data, &len, in->data, in->size), S2N_ERR_ENCRYPT);
 
-    S2N_ERROR_IF(len != in->size, S2N_ERR_ENCRYPT);
+    S2N_ERROR_IF(len != (int)in->size, S2N_ERR_ENCRYPT);
 
     return 0;
 }
@@ -42,10 +42,10 @@ static int s2n_stream_cipher_rc4_decrypt(struct s2n_session_key *key, struct s2n
 {
     gte_check(out->size, in->size);
 
-    int len = out->size;
+    int len = (int)out->size;
     GUARD_OSSL(EVP_DecryptUpdate(key->evp_cipher_ctx, out->data, &len, in->data, in->size), S2N_ERR_ENCRYPT);
 
-    S2N_ERROR_IF(len != in->size, S2N_ERR_ENCRYPT);
+    S2N_ERROR_IF(len != (int)in->size, S2N_ERR_ENCRYPT);
 
     return 0;
 }
