+++ tests/unit/s2n_ecc_evp_test.c
@@ -28,7 +28,7 @@ int main(int argc, char **argv) {
     BEGIN_TEST();
     {
         /* Test generate ephemeral keys for all supported curves */
-        for (int i = 0; i < s2n_all_supported_curves_list_len; i++) {
+        for (size_t i = 0; i < s2n_all_supported_curves_list_len; i++) {
             struct s2n_ecc_evp_params evp_params = {0};
             /* Server generates a key */
             evp_params.negotiated_curve = s2n_all_supported_curves_list[i];
@@ -38,7 +38,7 @@ int main(int argc, char **argv) {
     }
     {
         /* Test failure case for generate ephemeral key  when the negotiated curve is not set */
-        for (int i = 0; i < s2n_all_supported_curves_list_len; i++) {
+        for (size_t i = 0; i < s2n_all_supported_curves_list_len; i++) {
             struct s2n_ecc_evp_params evp_params = {0};
             /* Server generates a key */
             evp_params.negotiated_curve = NULL;
@@ -48,7 +48,7 @@ int main(int argc, char **argv) {
     }
     {
         /* Test generate ephemeral key and compute shared key for all supported curves */
-        for (int i = 0; i < s2n_all_supported_curves_list_len; i++) {
+        for (size_t i = 0; i < s2n_all_supported_curves_list_len; i++) {
             struct s2n_ecc_evp_params server_params = {0};
             struct s2n_ecc_evp_params client_params = {0};
             struct s2n_blob server_shared = {0};
@@ -85,8 +85,8 @@ int main(int argc, char **argv) {
     {
         /* Test failure case for computing shared key for all supported curves when the server
         and client curves donot match */
-        for (int i = 0; i < s2n_all_supported_curves_list_len; i++) {
-            for (int j = 0; j < s2n_all_supported_curves_list_len; j++) {
+        for (size_t i = 0; i < s2n_all_supported_curves_list_len; i++) {
+            for (size_t j = 0; j < s2n_all_supported_curves_list_len; j++) {
                 struct s2n_ecc_evp_params server_params = {0};
                 struct s2n_ecc_evp_params client_params = {0};
                 struct s2n_blob server_shared = {0};
@@ -120,7 +120,7 @@ int main(int argc, char **argv) {
     }
     {
         /* Test s2n_ecc_evp_write_params_point for all supported curves */
-        for (int i = 0; i < s2n_all_supported_curves_list_len; i++) {
+        for (size_t i = 0; i < s2n_all_supported_curves_list_len; i++) {
             struct s2n_ecc_evp_params test_params = {0};
             struct s2n_stuffer wire;
             uint8_t legacy_form;
@@ -151,7 +151,7 @@ int main(int argc, char **argv) {
     }
     {
         /* TEST s2n_ecc_evp_read_params_point for all supported curves */
-        for (int i = 0; i < s2n_all_supported_curves_list_len; i++) {
+        for (size_t i = 0; i < s2n_all_supported_curves_list_len; i++) {
             struct s2n_ecc_evp_params write_params = {0};
             struct s2n_blob point_blob;
             struct s2n_stuffer wire;
@@ -177,7 +177,7 @@ int main(int argc, char **argv) {
     }
     {
         /* TEST s2n_ecc_evp_parse_params_point for all supported curves */
-        for (int i = 0; i < s2n_all_supported_curves_list_len; i++) {
+        for (size_t i = 0; i < s2n_all_supported_curves_list_len; i++) {
             struct s2n_ecc_evp_params write_params = {0};
             struct s2n_ecc_evp_params read_params = {0};
             struct s2n_blob point_blob;
@@ -207,7 +207,7 @@ int main(int argc, char **argv) {
     }
     {
         /* Test read/write/parse params for all supported curves */
-        for (int i = 0; i < s2n_all_supported_curves_list_len; i++) {
+        for (size_t i = 0; i < s2n_all_supported_curves_list_len; i++) {
             struct s2n_ecc_evp_params write_params = {0};
             struct s2n_ecc_evp_params read_params = {0};
             struct s2n_stuffer wire;
@@ -240,7 +240,7 @@ int main(int argc, char **argv) {
     }
     {
         /* Test generate/read/write/parse and compute shared secrets for all supported curves */
-        for (int i = 0; i < s2n_all_supported_curves_list_len; i++) {
+        for (size_t i = 0; i < s2n_all_supported_curves_list_len; i++) {
             struct s2n_ecc_evp_params server_params = {0};
             struct s2n_ecc_evp_params read_params = {0};
             struct s2n_ecc_evp_params client_params = {0}; 
@@ -294,7 +294,7 @@ int main(int argc, char **argv) {
     }
     {
         /* Test generate->write->read->compute_shared with all supported curves */
-    for (int i = 0; i < s2n_all_supported_curves_list_len; i++) {
+    for (size_t i = 0; i < s2n_all_supported_curves_list_len; i++) {
             struct s2n_ecc_evp_params server_params = {0}, client_params = {0};
             struct s2n_stuffer wire;
             struct s2n_blob server_shared, client_shared, ecdh_params_sent, ecdh_params_received;
